allow --env-file flag multiple times
This change updates to the --env-file flag definition so that it can be
repeated multiple times. This is useful in settings where benthos is
supplied with multiple files containing environment variables. For
example, a container orchestration framework can inject one file
containing application secrets, another file containing database access
credentials and there may be a file included in the repository of a
benthos project. These can be loaded like so:

```
benthos \
  --env-file ./.env.production \
  --env-file /tmpfs/secrets.env \
  --env-file /tmpfs/db.env \
  --config config.yml
```