rewrite gcp_pubsub output
This change rewrites the gcp_pubsub output to use the new config API and
to lean on the pubsub client's internal batching mechanism. This is
referred to as the "internal buffer" in docs to distinguish it from
Benthos' batching mechanism. The rewrite also benefits from added tests
and a pubsub mock.

Signed-off-by: Georges Haidar <ghaidar0@gmail.com>