use SortGroup's batch as the basis for BatchError
This change maintains a sort group and associated batch as the basis of
service.BatchError. Creating a sort group returns a batch with message
parts that are shallow clones of the input batch's messages with a
context value tagging them in the sort group. Yet when calling
BatchError.WalkMessages we use the original batch with the sort group
and when it tries to locate messages by their reference it always fails
resulting in no iteration.